requirements.txt
fastapi
uvicorn
python-docx
gtts
moviepy
pydub
